feat: 로직 위젯으로 이동
